requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
anyio==3.7.1
astral==2.2
async-timeout==4.0.0
atomicwrites==1.4.0
attrs==21.2.0
awesomeversion==21.11.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==2.0.12
ciso8601==2.2.0
cryptography==35.0.0
exceptiongroup==1.2.2
frozenlist==1.4.1
h11==0.12.0
homeassistant==2021.12.10
httpcore==0.14.7
httpx==0.21.0
idna==3.7
ifaddr==0.1.7
Jinja2==3.0.3
MarkupSafe==2.1.5
multidict==6.0.5
paho-mqtt==1.6.1
pycparser==2.22
PyJWT==2.1.0
python-slugify==4.0.1
pytz==2024.1
PyYAML==6.0
requests==2.26.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
stringcase==1.2.0
text-unidecode==1.3
typing-extensions==4.12.2
urllib3==1.26.19
voluptuous==0.12.2
voluptuous-serialize==2.5.0
yarl==1.6.3
